rename module files from core to module<name> to remove clash with rust core
